Move Curve25519 operations into their own module.
This supports both the short-term oxidization of Curve25519 operations
and the longer-term unification with the NIST EC code, whose module
structure we now mirror more closely.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.
